Reverted to previous folder structure.
