+++ src/components/Heading.js
@@ -2,7 +2,7 @@ import React, {useState} from 'react';
 import 'antd/dist/antd.css'
 import {Menu, Button, Row, Col} from 'antd';
 import styled from 'styled-components';
-import Filter from '../dashboard/Filter'
+import Filter from './Filter'
 const { SubMenu } = Menu;
